% Auto-generated: do not edit by hand
\name{beforeAfter}

\alias{beforeAfter}

\title{BeforeAfter component}

\description{
A light wrapper of BeforeAfterSlider.
}

\usage{
beforeAfter(children=NULL, id=NULL, after=NULL, afterClassName=NULL,
afterProps=NULL, before=NULL, beforeClassName=NULL,
beforeProps=NULL, className=NULL, defaultProgress=NULL,
height=NULL, width=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{after}{Character. }

\item{afterClassName}{Character. }

\item{afterProps}{Named list. }

\item{before}{Character. }

\item{beforeClassName}{Character. }

\item{beforeProps}{Named list. }

\item{className}{Character. The class of the component}

\item{defaultProgress}{Numeric. }

\item{height}{Numeric. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

