% Auto-generated: do not edit by hand
\name{lottie}

\alias{lottie}

\title{Lottie component}

\description{

}

\usage{
lottie(children=NULL, id=NULL, ariaLabel=NULL, ariaRole=NULL,
className=NULL, direction=NULL, height=NULL,
isClickToPauseDisabled=NULL, isPaused=NULL, isStopped=NULL,
options=NULL, segments=NULL, speed=NULL, style=NULL,
title=NULL, url=NULL, width=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{ariaLabel}{Character. }

\item{ariaRole}{Character. }

\item{className}{Character. The class of the component}

\item{direction}{Numeric. }

\item{height}{Character. Pixel value for containers height.}

\item{isClickToPauseDisabled}{Logical. }

\item{isPaused}{Logical. }

\item{isStopped}{Logical. }

\item{options}{Named list. Options passed to the Lottie animation (see https://www.npmjs.com/package/react-lottie for details)}

\item{segments}{List of numerics. }

\item{speed}{Numeric. }

\item{style}{Character. }

\item{title}{Character. }

\item{url}{Character. If set, data will be downloaded from this url.}

\item{width}{Character. Pixel value for containers width.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

