% Auto-generated: do not edit by hand
\name{mermaid}

\alias{mermaid}

\title{Mermaid component}

\description{

}

\usage{
mermaid(id=NULL, chart=NULL, className=NULL, config=NULL, name=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{chart}{Character. The mermaid code of your chart. Check Mermaid js documentation for details}

\item{className}{Character. The class of the component}

\item{config}{Named list. On optional object with one of several Mermaid config parameters. Check Mermaid js documentation for details}

\item{name}{Character. On optional name of your mermaid diagram/flowchart/gantt etc.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

