% Auto-generated: do not edit by hand
\name{ticker}

\alias{ticker}

\title{Ticker component}

\description{
A light wrapper of ReactTicker.
}

\usage{
ticker(children=NULL, id=NULL, className=NULL, direction=NULL,
height=NULL, mode=NULL, move=NULL, offset=NULL, speed=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{className}{Character. The class of the component}

\item{direction}{A value equal to: "toright", "toleft". }

\item{height}{Numeric | character. }

\item{mode}{A value equal to: "chain", "await", "smooth". }

\item{move}{Logical. }

\item{offset}{Numeric | character. }

\item{speed}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

